Move sqlitetuplebatchiterator into its own class
